Remove wrapping commit hashes in single quotes
This attempts to fix the issue of git diff failing on msysgit when
wrapping commit hashes in quotes.

For example running:
  git diff '1dee4678' 'e7156d09'
    results in:
      fatal: ambiguous argument ''1dee4678'': unknown revision or path
      not in the working tree.
whereas:
  git diff 1dee4678 e7156d09
    works as expected!

This happens probably because in windows quotes are used to define
system paths.

This change attempts to fix this issue by removing surrounding single
quotes in --pretty-format template.