Now you can set your own config file (an object/JSON) on the constructor of DPLOY if you don't want to use the YAML file.
t task (#24), where you can set the options directly on your
Gruntfile.js.

It's fairly simple to use:

```
var DPLOY = require("dploy");
var settings = {
    host: 'ftp.example.com',
    user: 'user',
    pass: 'password',
    path: {
        remote: 'public_html/',
        local: ''
    }
}

new DPLOY(settings, function() {
    console.log("Upload completed");
});
```

Where the first parameter is the config file for that server and the
second one is a callback for when the upload is completed. Both are
optional, of course.