Added more information about the new version
The changes are now documented on the CHANGELOG file and the
README file is also updated.
Some other bits to help you debug errors were added too - now if
an error occurs while you are deploying, a new `dploy.log` will be
created at the root of your repository.

The SFTP was also added, but not fully tested yet.